MoveProver

+++ action/action_move_prover.go
@@ -110,6 +110,9 @@ func (a *MoveProverAction) Hook() (*model.ActionResult, error) {
 		replaceAfterString := compile.ReplaceAllString(out, "")
 		replaceAfterString = strings.ReplaceAll(replaceAfterString, "\u001B", "")
 		//replaceAfterString = strings.ReplaceAll(replaceAfterString, "[38;5;11m", "")
+		if strings.Contains(replaceAfterString, "Error") && strings.Contains(replaceAfterString, "failed") {
+			return nil, errors.New(replaceAfterString)
+		}
 		create, err := os.Create(dest)
 		if err != nil {
 			return nil, err
